some fixes

+++ routes/offer_route.py
@@ -2,6 +2,7 @@
 from flask import Blueprint, jsonify, request
 
 from models.offer import Offer
+from blueprints.admin import admin_bp
 from blueprints.user import user_bp
 
 
@@ -10,10 +11,10 @@ from blueprints.user import user_bp
 def get_all_offers():
     offers = Offer.get_all_offers()
     if offers :
-        res = jsonify({"message" : 'Get request succeeded'  , 'data': offers})
+        res = jsonify({"Message" : 'Get request succeeded'  , 'data': offers})
         res.status_code = 200
     else:
-        res = jsonify({'message': 'Unable to get all offers'})
+        res = jsonify({'Error': 'Unable to get all offers'})
         res.status_code = 404
     return res
     
@@ -24,10 +25,10 @@ def filter_offers():
     data = request.get_json()
     offers = Offer.filter_offers(data)
     if offers :
-        res = jsonify({"message" : 'Get request succeeded'  , 'data': offers})
+        res = jsonify({"Message" : 'Get request succeeded'  , 'data': offers})
         res.status_code = 200
     else:
-        res = jsonify({'message': 'Unable to get all offers'})
+        res = jsonify({'Error': 'Unable to get all offers'})
         res.status_code = 404
     return res
 
@@ -35,10 +36,10 @@ def filter_offers():
 def get_one_offer(_id):
     offer = Offer.get_one_offer(_id)
     if offer :
-        res = jsonify({"message" : 'Get request succeeded' ,'data': offer})
+        res = jsonify({"Message" : 'Get request succeeded' ,'data': offer})
         res.status_code = 200
     else:
-        res = jsonify({'message': 'Unable to get offer'})
+        res = jsonify({'Error': 'Unable to get offer'})
         res.status_code = 404
     return res
 
@@ -49,34 +50,48 @@ def create_offer():
     print(data)
     result = Offer.create_offer(data['quantity'], data['quality'], data['priceUnit'], data['unit'], data['state'], data['actorType'], data['actorRef'] )
     if result :
-        res = jsonify({'message': 'offer created'})
+        res = jsonify({'Message': 'offer created'})
         res.status_code = 200
     else:
-        res = jsonify({'message': 'Unable to get offer'})
+        res = jsonify({'Error': 'Unable to get offer'})
         res.status_code = 404
     return res
 
 
 @user_bp.route('/offers/<string:_id>', methods=['PUT'])
 def update_offer(_id):
+    auth_header = request.headers.get('Authorization')
+    jwt_token = auth_header.split(' ')[1]
+    decoded_token = jwt.decode(jwt_token, SECRET_KEY, algorithms=['HS256'])
+    if  decoded_token['user_id'] != _id:
+        res = jsonify({'Error' : "Unauthorized"})
+        res.status_code = 401
+        abort(res)
     data = request.get_json()
-    result = Offer.update_offer(_id, data['quantity'], data['quality'], data['priceUnit'], data['unit'],data['state'], data['actorType'], data['actorRef'])
+    result = Offer.update_offer(_id, data)
     if result :
-        res = jsonify({'message': 'offer updated'})
+        res = jsonify({'Message': 'offer updated'})
         res.status_code = 200
     else:
-        res = jsonify({'message': 'Unable to update offer'})
+        res = jsonify({'Error': 'Unable to update offer'})
         res.status_code = 404
     return res
 
 
 @user_bp.route('/offers/<string:_id>', methods=['DELETE'])
 def delete_offer(_id):
+    auth_header = request.headers.get('Authorization')
+    jwt_token = auth_header.split(' ')[1]
+    decoded_token = jwt.decode(jwt_token, SECRET_KEY, algorithms=['HS256'])
+    if  decoded_token['user_id'] != _id:
+        res = jsonify({'Error' : "Unauthorized"})
+        res.status_code = 401
+        abort(res)
     result = Offer.delete_offer(_id)
     if result > 0:
-        res = jsonify({'message': 'offer deleted'})
+        res = jsonify({'Message': 'offer deleted'})
         res.status_code = 200
     else:
-        res = jsonify({'message': 'Unable to delete offer'})
+        res = jsonify({'Error': 'Unable to delete offer'})
         res.status_code = 404
     return res
